Move math types to new location
