+++ ros_bt_py/ros_bt_py/helpers.py
@@ -106,25 +106,6 @@ def json_decode(data):
     return jsonpickle.decode(data)
 
 
-class MathUnaryOperator(object):
-    def __init__(self, operator):
-        self.operator = operator
-
-
-class MathBinaryOperator(object):
-    def __init__(self, operator):
-        self.operator = operator
-
-
-class MathOperandType(object):
-    def __init__(self, operand_type):
-        self.operand_type = operand_type
-
-
-class MathUnaryOperandType(object):
-    def __init__(self, operand_type):
-        self.operand_type = operand_type
-
 
 class HashableCapabilityInterface:
     """Wrapper class to allow for the hashing of capability interfaces."""
